Break out separate module for data generation
